+++ lib/player.h
@@ -11,8 +11,20 @@ typedef enum {
 	ADDPLAYER_DB_ERROR			// Database error
 } AddPlayerReturn;
 
+typedef enum {
+	RENAMEPLAYER_OK = 0,			// Success
+	RENAMEPLAYER_EXISTS,			// Player with that name already exists
+	RENAMEPLAYER_INVALID_NAME,		// Invalid name provided
+	RENAMEPLAYER_NO_NAME,			// No name provided
+	RENAMEPLAYER_VALIDATION_ERROR,	// Regex error
+	RENAMEPLAYER_DB_ERROR			// Database error
+} RenamePlayerReturn;
+
 int add_player(const char* name);
 int delete_player();
-int rename_player();
+int rename_player(
+		const char* current_name,
+		const char* new_name
+		);
 
 #endif